backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
opencv-python-headless
tensorflow-cpu